pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = "Todo List with Async SQLAlchmey and MariaDB"
authors = ["Vitor Nere <nere.swe@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
fastapi = "^0.85.1"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
python-decouple = "^3.6"
fastapi-pagination = "^0.10.0"
newrelic = "^8.2.1"
python-json-logger = "^2.0.4"
alembic = "^1.8.1"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.41"}
asyncmy = "^0.2.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.19.0"
pytest-mock = "^3.10.0"
ipython = "^8.5.0"
pre-commit = "^2.20.0"
pylint-pydantic = "^0.1.4"
mypy = "^0.982"
sqlalchemy2-stubs = "^0.0.2a29"
debugpy = "^1.6.3"
asgi-lifespan = "^1.0.1"
httpx = "^0.23.0"
pytest-env = "^0.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
